# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NanoPay API
  version: "2.0.0"
  description: |
    API Backend NanoPay — commande centrale unique reliant Client, Commerçant, Banque et Admin,
    avec trois modes de financement : **Épargne**, **Coffre** (prélèvement sur salaire) et **Crédit**.

    Conforme au « Cahier de construction BACKEND ».

    ## Contrat de connexion
    | Élément | Contrat |
    |---|---|
    | Base URL | `/api` (alias historique `/api/v1`) |
    | Format | JSON UTF-8 (multipart pour les documents) |
    | Auth | `Authorization: Bearer <accessToken>` + refresh token (`/auth/refresh`, rotation) |
    | Succès | `{ "data": ..., "meta"?: ... }` |
    | Erreur | `{ "code", "message", "details"?, "requestId" }` (`X-Request-Id` aussi en en-tête) |
    | Pagination | query `page`, `limit` ; `meta` = `{ page, limit, total, nextPage }` |
    | Dates | ISO 8601 |
    | Montants | FCFA **entiers**, jamais de flottant |
    | Identifiants | UUID serveur ; `orderNumber` lisible (ex. `NP-2026-00001245`) pour l'UI |
    | Temps réel | pas de WebSocket : polling contrôlé de `GET /notifications` et des ressources |

    Les règles financières et de sécurité sont **toujours** appliquées côté backend ; le frontend
    ne fait que consommer ces contrats.

    ## Circuit
    Client → produit → commande (`CREEE`) → `select-financing` (`FINANCEMENT_EN_COURS`) →
    Épargne / Coffre / Crédit → `FINANCEE` → `PRETE_A_LIVRER` (+ QR) → scan commerçant →
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /orders/{id}/savings:
    post:
      tags: [Savings]
      summary: Ouvre le plan d'épargne de la commande
      description: |
        Versement minimum / durée max selon le prix : ≤50 000 → 1 000 FCFA, 8 mois ; 50 000–100 000 → 2 000, 18 mois ;
        >100 000 → 5 000, 36 mois (paramétrable). La date de départ est celle du **premier versement validé**.
        `paymentReference` = `orderId` à fournir lors du versement Mobile Money.
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        content: { application/json: { schema: { type: object, properties: { durationMonths: { type: integer, minimum: 1 } } } } }
      responses:
        "201": { description: Épargne créée }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /savings/{id}:
    get:
      tags: [Savings]
      summary: Détail (progression, reste, jours restants)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Épargne }
  /savings/{id}/history:
    get:
      tags: [Savings]
      summary: Historique des paiements de l'épargne
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Page de paiements }
  /savings/{id}/extension:
    post:
      tags: [Savings]
      summary: Prolonge l'échéance (2 mois maximum au total, avant l'échéance)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [months], properties: { months: { type: integer, minimum: 1, maximum: 2 } } } } }
      responses:
        "200": { description: Épargne en `PROLONGATION` }
        "400": { $ref: "#/components/responses/Error" }
  # ------------------------------------------------------------------ VAULT
  /orders/{id}/vault:
    post:
      tags: [Vault]
      summary: Programme le Coffre (prélèvement automatique sur salaire)
      description: "Réservé aux salariés dont le profil crédit est déjà validé. `multipart` : `durationMonths` + `salaryDebitAuthorization` (optionnel)."
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [durationMonths]
              properties:
                durationMonths: { type: integer }
                salaryDebitAuthorization: { type: string, format: binary }
      responses:
        "201": { description: Coffre créé }
        "403": { $ref: "#/components/responses/Error" }
  /vaults/{id}:
    get:
      tags: [Vault]
      summary: Détail du Coffre
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Coffre }
  /vaults/{id}/use:
    post:
      tags: [Vault]
      summary: Utilise un Coffre complet (`OBJECTIF_ATTEINT`) → commande prête + QR
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: "`data` = { orderId, status, qrToken, expiresAt }" }
        "409": { $ref: "#/components/responses/Error" }
  /vaults/{id}/switch-to-credit:
    post:
      tags: [Vault]
      summary: Bascule vers un crédit pour le reste à financer
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [durationMonths], properties: { durationMonths: { type: integer } } } } }
      responses:
        "201": { description: Demande de crédit créée, Coffre `BASCULE_VERS_CREDIT` }
        "403": { $ref: "#/components/responses/Error" }
  # ------------------------------------------------------------------ CREDIT
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    Error:
      description: "Erreur — `{ code, message, details?, requestId }`"
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
...
    Error:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        details: {}
        requestId: { type: string }
...